pyproject.toml
[project]
name = "Slyth"
version = "0.1.0"
description = "Multiplayer snake game"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==25.1.0",
    "loguru==0.7.3",
    "pyyaml==6.0.3",
    "websockets==15.0.1",
]

[project.scripts]
slyth-server = "src.server.server_run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
